Move DeleteButton into it's own component
