+++ src/components/PlantList.js
@@ -9,10 +9,7 @@ function PlantList({ plants, setPlants, onSearchForPlants, onHandleDelete }) {
       {renderPlantList.map((plant) => (
         <PlantCard
           key={plant.id}
-          id={plant.id}
-          name={plant.name}
-          image={plant.image}
-          price={plant.price}
+          plant={plant}
           setPlants={setPlants}
           plants={plants}
           onDeleteItem={onHandleDelete}
